requirements.txt
streamlit>=1.28.0
openai>=1.3.0
PyMuPDF>=1.23.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0